feat(react-icons): revert to HOC, use common build system
