+++ packages/react-core/src/components/Progress/ProgressContainer.tsx
@@ -34,7 +34,7 @@ export interface ProgressContainerProps extends Omit<React.HTMLProps<HTMLDivElem
   value: number;
 }
 
-const variantToIcon: { [k: string]: React.FunctionComponent } = {
+const variantToIcon = {
   danger: TimesCircleIcon,
   success: CheckCircleIcon
 };
