--- packages/react-icons/scripts/icons/index.js
@@ -1,9 +1,0 @@
-const faIcons = require('./fontawesomeIcons');
-const patternflyIcons = require('./patternflyIcons');
-const customIcons = require('./customIcons');
-
-module.exports = {
-  ...faIcons,
-  ...patternflyIcons,
-  ...customIcons
-};
